Create collabFiltering.py
The hyperparameters need to be tweaked, the results are currently very not good.  Unless you like a negative r2 score (meaning this is impossibly bad), in which case this model is perfect

Tweaking the model architecture, learning rate, batch size, and number of epochs.  Also could adjust the activation functions, I have not tried that.  Seems to work best with smaller hidden layers